Remove user visible strings from the javascript file
Where I create the links for the tag-cloud and archive, I was setting
a title with a fixed string in the javascript file. But that's inflexible
if you need to translate the interface. Now, use the title attribute of
the parent UL element as a template.

I don't do real templating for now, but that may happen later
(like Python style .format).