Implemented pjax
https://github.com/defunkt/jquery-pjax

With this commit there are several changes.
I now use the `jquery.pjax.js` library so the Makefile and index.html
reflect this addtion.

`script/paste.js` is changed so that it calls $.pjax instead of redirecting
The `paste.html` template is split with a 'partial' part for the paste content